[test] Polish refactor
