+++ packages/grid/_modules_/grid/utils/__tests__/classnames.tests.ts
@@ -5,15 +5,18 @@ describe('utils: classnames', () => {
     const css = classnames('hello', 'world');
     expect(css).toBe('hello world');
   });
+
   it('should work with css rules object', () => {
     const css = classnames('working', { yes: true, no: false });
     expect(css).toBe('working yes');
   });
-  it('should work with string arrays ', () => {
+
+  it('should work with string arrays', () => {
     const css = classnames(['working', 'from', 'home']);
     expect(css).toBe('working from home');
   });
-  it('should work with all types of params mixed together ', () => {
+
+  it('should work with all types of params mixed together', () => {
     const css = classnames(
       'people',
       ['working', 'from', 'home'],
