+++ packages/grid/data-grid/src/tests/toolbar.DataGrid.test.tsx
@@ -3,7 +3,7 @@ import {
   createClientRenderStrictMode,
   // @ts-expect-error need to migrate helpers to TypeScript
   fireEvent,
-} from 'test/utils/index';
+} from 'test/utils';
 import { expect } from 'chai';
 import { DataGrid } from '@material-ui/data-grid';
 import {
