+++ packages/grid/x-grid/src/tests/layout.XGrid.test.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { createClientRenderStrictMode } from 'test/utils';
-import { raf } from 'test/utils/helperFn';
 import { expect } from 'chai';
 import { XGrid } from '@material-ui/x-grid';
 
@@ -63,29 +62,5 @@ describe('<XGrid /> - Layout', () => {
         container.firstChild.firstChild.firstChild,
       );
     });
-  });
-
-  it('should resize the width of the columns', async () => {
-    interface TestCaseProps {
-      width?: number;
-    }
-    const TestCase = (props: TestCaseProps) => {
-      const { width = 300 } = props;
-      return (
-        <div style={{ width, height: 300 }}>
-          <XGrid {...baselineProps} />
-        </div>
-      );
-    };
-
-    const { container, setProps } = render(<TestCase width={300} />);
-    let rect;
-    rect = container.querySelector('[role="row"][data-rowindex="0"]').getBoundingClientRect();
-    expect(rect.width).to.equal(300 - 2);
-
-    setProps({ width: 400 });
-    await raf(); // wait for the AutoSize's dimension detection logic
-    rect = container.querySelector('[role="row"][data-rowindex="0"]').getBoundingClientRect();
-    expect(rect.width).to.equal(400 - 2);
   });
 });
