+++ packages/grid/x-grid/src/tests/pagination.XGrid.test.tsx
@@ -1,9 +1,16 @@
-import { act, render } from '@testing-library/react';
+import {
+  createClientRenderStrictMode,
+  // @ts-expect-error need to migrate helpers to TypeScript
+  act,
+} from 'test/utils';
 import * as React from 'react';
 import { expect } from 'chai';
 import { XGrid, useApiRef } from '@material-ui/x-grid';
 
 describe('<XGrid /> - Pagination', () => {
+  // TODO v5: replace with createClientRender
+  const render = createClientRenderStrictMode();
+
   const baselineProps = {
     rows: [
       {
@@ -29,7 +36,7 @@ describe('<XGrid /> - Pagination', () => {
     }
   });
 
-  it('should apply setPage correctly', async () => {
+  it('should apply setPage correctly', () => {
     let apiRef;
     const GridTest = () => {
       apiRef = useApiRef();
@@ -41,7 +48,8 @@ describe('<XGrid /> - Pagination', () => {
       );
     };
 
-    render(<GridTest />);
+    // TODO fix StrictMode support
+    render(<GridTest />, { strict: false });
 
     let cell = document.querySelector('[role="cell"][aria-colindex="0"]')!;
     expect(cell).to.have.text('Nike');
